+++ eras/babbage/impl/testlib/Test/Cardano/Ledger/Babbage/ImpTest.hs
@@ -7,7 +7,9 @@
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
-module Test.Cardano.Ledger.Babbage.ImpTest () where
+module Test.Cardano.Ledger.Babbage.ImpTest (
+  module Test.Cardano.Ledger.Alonzo.ImpTest,
+) where
 
 import Cardano.Crypto.DSIGN (DSIGNAlgorithm (..), Ed25519DSIGN)
 import Cardano.Crypto.Hash (Hash)
@@ -15,16 +17,9 @@ import Cardano.Ledger.Babbage (BabbageEra)
 import Cardano.Ledger.Core
 import Cardano.Ledger.Crypto (Crypto (..))
 import Cardano.Ledger.Plutus.Language (SLanguage (..))
-import Test.Cardano.Ledger.Allegra.ImpTest (impAllegraSatisfyNativeScript)
-import Test.Cardano.Ledger.Alonzo.ImpTest (
-  AlonzoEraImp (..),
-  alonzoFixupTx,
-  initAlonzoImpNES,
-  plutusTestScripts,
- )
+import Test.Cardano.Ledger.Alonzo.ImpTest
 import Test.Cardano.Ledger.Babbage.TreeDiff ()
 import Test.Cardano.Ledger.Common
-import Test.Cardano.Ledger.Shelley.ImpTest (ShelleyEraImp (..))
 
 instance
   ( Crypto c
@@ -38,6 +33,8 @@ instance
   initImpNES = initAlonzoImpNES
   impSatisfyNativeScript = impAllegraSatisfyNativeScript
   fixupTx = alonzoFixupTx
+
+instance ShelleyEraImp (BabbageEra c) => MaryEraImp (BabbageEra c)
 
 instance ShelleyEraImp (BabbageEra c) => AlonzoEraImp (BabbageEra c) where
   scriptTestContexts = plutusTestScripts SPlutusV1 <> plutusTestScripts SPlutusV2